Initial version of a POC for GraphRAG
This is based on the Microsoft project https://microsoft.github.io/graphrag/. This is not yet a fully implemented solution, but it has a working graph creation pipeline. To be done is accurate summarization on top of communities to persist semantic information about the overall knowledge base structure. From there, we'll add a layer to query and retrieve information.